// src/generated/models/Cr108_requestsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr108_requestscr108_status = {
  1: 'Approved',
  2: 'Rejected',
  3: 'MoreInfoRequestedFromUser',
  4: 'Pending'
} as const;
export type Cr108_requestscr108_status = keyof typeof Cr108_requestscr108_status;
export const Cr108_requestsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr108_requestsstatecode = keyof typeof Cr108_requestsstatecode;
export const Cr108_requestsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr108_requestsstatuscode = keyof typeof Cr108_requestsstatuscode;

export interface Cr108_requestsBase {
  "cr108_approver_admin@odata.bind"?: string;
  cr108_date_resolved?: string;
  cr108_justification?: string;
  cr108_key: string;
  cr108_requester?: string;
  cr108_requestid: string;
  cr108_shared_id?: number;
  "cr108_software@odata.bind": string;
  cr108_status: Cr108_requestscr108_status;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr108_requestsstatecode;
  statuscode?: Cr108_requestsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr108_requests extends Cr108_requestsBase {
  cr108_approver_adminname?: string;
  cr108_approver_adminyominame?: string;
  cr108_softwarename?: string;
  cr108_statusname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  cr108_approver_admin?: object;
  _cr108_approver_admin_value?: string;
  cr108_software?: object;
  _cr108_software_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}